+++ finagle-core/src/test/scala/com/twitter/finagle/context/MarshalledContextTest.scala
@@ -21,6 +21,134 @@ class MarshalledContextTest extends AbstractContextTest {
     }
   }
 
+  def withCaseInsensitiveOverride[R](enabled: Boolean)(f: => R): R = {
+    val fraction = if (enabled) 1.0 else 0.0
+    com.twitter.finagle.toggle.flag.overrides
+      .let("com.twitter.finagle.context.MarshalledContextLookupId", fraction) {
+        f
+      }
+  }
+
+  test("Key#marshalId is used when marshalling") {
+    val key = new ctx.Key[String]("C.kEy") {
+      def marshal(value: String) = Buf.Utf8(value)
+      def tryUnmarshal(buf: Buf) = buf match {
+        case Buf.Utf8(value) => Return(value)
+      }
+    }
+
+    ctx.let(key, "bar") {
+      assert(
+        ctx.marshal() == Map(
+          key.marshalId -> Buf.Utf8("bar")
+        )
+      )
+    }
+  }
+
+  test("When disabled, key lookups are case sensitive") {
+    withCaseInsensitiveOverride(false) {
+      val ctx = new MarshalledContext
+      val lowerKey = new ctx.Key[String]("lowerkey") {
+        def marshal(value: String) = Buf.Utf8(value)
+        def tryUnmarshal(buf: Buf) = buf match {
+          case Buf.Utf8(value) => Return(value)
+        }
+      }
+
+      val upperKey = new ctx.Key[String]("LOWERKEY") {
+        def marshal(value: String) = Buf.Utf8(value)
+        def tryUnmarshal(buf: Buf) = buf match {
+          case Buf.Utf8(value) => Return(value)
+        }
+      }
+
+      ctx.let(lowerKey, "bar") {
+        val lowerValue = ctx.get(lowerKey)
+        assert(lowerValue.isDefined)
+
+        val upperValue = ctx.get(upperKey)
+        assert(upperValue == None)
+      }
+    }
+  }
+
+  test("When enabled, keys unmarshal a into a case insensitive format") {
+    withCaseInsensitiveOverride(true) {
+      val ctx = new MarshalledContext
+      val lowerKey = new ctx.Key[String]("foo") {
+        def marshal(value: String) = Buf.Utf8(value)
+        def tryUnmarshal(buf: Buf) = buf match {
+          case Buf.Utf8(value) => Return(value)
+        }
+      }
+      val upperKey = new ctx.Key[String]("FOO") {
+        def marshal(value: String) = Buf.Utf8(value)
+        def tryUnmarshal(buf: Buf) = buf match {
+          case Buf.Utf8(value) => Return(value)
+        }
+      }
+
+      val data = ctx.let(upperKey, "bar") { ctx.marshal() }
+      ctx.letUnmarshal(data) {
+        val value = ctx.get(lowerKey)
+        assert(value.isDefined)
+        assert(value.get == "bar")
+      }
+    }
+  }
+
+  test("when enabled, key lookups are case insenstive") {
+    withCaseInsensitiveOverride(true) {
+      val ctx = new MarshalledContext
+      val lowerKey = new ctx.Key[String]("foo") {
+        def marshal(value: String) = Buf.Utf8(value)
+        def tryUnmarshal(buf: Buf) = buf match {
+          case Buf.Utf8(value) => Return(value)
+        }
+      }
+      val upperKey = new ctx.Key[String]("FOO") {
+        def marshal(value: String) = Buf.Utf8(value)
+        def tryUnmarshal(buf: Buf) = buf match {
+          case Buf.Utf8(value) => Return(value)
+        }
+      }
+
+      ctx.let(lowerKey, "hello") {
+        val value = ctx.get(upperKey)
+        assert(value.isDefined)
+        assert(value.get == "hello")
+
+        val originalKeyValue = ctx.get(lowerKey)
+        assert(originalKeyValue.isDefined)
+        assert(originalKeyValue.get == "hello")
+      }
+    }
+  }
+
+  test("Flipping toggle does not change existing context") {
+    val ctx = withCaseInsensitiveOverride(false) { new MarshalledContext }
+    val lowerKey = new ctx.Key[String]("lowerkey") {
+      def marshal(value: String) = Buf.Utf8(value)
+      def tryUnmarshal(buf: Buf) = buf match {
+        case Buf.Utf8(value) => Return(value)
+      }
+    }
+
+    val upperKey = new ctx.Key[String]("LOWERKEY") {
+      def marshal(value: String) = Buf.Utf8(value)
+      def tryUnmarshal(buf: Buf) = buf match {
+        case Buf.Utf8(value) => Return(value)
+      }
+    }
+
+    withCaseInsensitiveOverride(true) {
+      ctx.let(lowerKey, "foo") {
+        assert(ctx.get(upperKey) == None)
+      }
+    }
+  }
+
   test("Translucency: pass through, replace") {
     ctx.let(b, 333) {
       ctx.letUnmarshal(Seq(Buf.Utf8("bleep") -> Buf.Utf8("bloop"))) {
@@ -88,7 +216,7 @@ class MarshalledContextTest extends AbstractContextTest {
       val roundTrip = ctx.doUnmarshal(Map.empty, ctx.marshal())
 
       def checkKey(key: ctx.Key[_]): Unit = {
-        roundTrip(key.marshalId) match {
+        roundTrip(key.lookupId) match {
           case t: ctx.Translucent => assert(t.unmarshal(key) == ctx.get(key))
           case other => fail(s"Unexpected structure: $other")
         }